requirements.txt
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
branca==0.6.0
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.1
django-cors-headers==4.3.1
django-jazzmin==3.0.0
djangorestframework==3.14.0
docopt==0.6.2
drf-yasg==1.21.7
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
folium==0.14.0
identify==2.5.33
idna==3.4
inflection==0.5.1
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nodeenv==1.8.0
numpy==1.26.3
opencv-python==4.9.0.80
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.1.0
pre-commit==3.6.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyfiglet==1.0.2
Pygments==2.18.0
pyTelegramBotAPI==4.14.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4
traitlets==5.14.3
uritemplate==4.1.1
urllib3==2.0.7
uuid==1.30
virtualenv==20.25.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
